Remove create component, move logic to main create page for now
